Simplify action class organisation
