--- src/main/kotlin/com/fwdekker/randomness/decimal/DecimalSettingsAction.kt
@@ -1,14 +1,0 @@
-package com.fwdekker.randomness.decimal
-
-import com.fwdekker.randomness.SettingsAction
-
-
-/**
- * Controller for random decimal generation settings.
- */
-class DecimalSettingsAction : SettingsAction() {
-    override val title = "Decimal Settings"
-
-
-    public override fun createDialog() = DecimalSettingsDialog()
-}
