--- src/main/kotlin/com/fwdekker/randomness/string/StringGroupAction.kt
@@ -1,13 +1,0 @@
-package com.fwdekker.randomness.string
-
-import com.fwdekker.randomness.DataGroupAction
-
-
-/**
- * All actions related to inserting strings.
- */
-class StringGroupAction : DataGroupAction() {
-    override val insertAction = StringInsertAction()
-    override val insertArrayAction = StringInsertAction().ArrayAction()
-    override val settingsAction = StringSettingsAction()
-}
